reporting system finished (including message room)

+++ backend/chatSystem/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
-from Calcura.models import MessageRoom, Calculator
+from Calcura.models import MessageRoom, Calculator, MessageReport
 from Calcura.views import generateId
 from django.http import HttpResponseRedirect
 from .models import Messages
@@ -42,9 +42,15 @@ def chatPage(request,roomId): #https://www.youtube.com/watch?v=F4nwRQPXD8w&ab_ch
         for message in Messages.objects.all():
             print(message.id)
         if request.method=="POST":
-            id=request.POST["message"]
-            if Messages.objects.filter(id=id).exists():
-                Messages.objects.get(id=id).delete()
+            if "report" in request.POST:
+                room = MessageRoom.objects.get(id=request.POST['roomid'])
+                description = request.POST['description']
+                r = MessageReport(room=room, description=description, reporter=request.user, reported=otherUser)
+                r.save()
+            else:
+                id=request.POST["message"]
+                if Messages.objects.filter(id=id).exists():
+                    Messages.objects.get(id=id).delete()
 
 
     else:
